worker/requirements.txt
backoff==2.0.1
pika==1.2.1
psycopg2-binary==2.8.6
python-dotenv==0.20.0
loguru==0.6.0
